Cloudformation support for Secrets Manager Integration (#26)
The changes provide Cloudformation support for Secrets Manager integration with Redshift Serverless. With this new feature, customers can opt in to store their namespace's admin credentials in a service linked secret in Secrets Manager. It allows us to create/modify Redshift Serverless namespaces with Secrets Manager support using the Cloudformation template. The changes in this request allow us to use create-namespace and update-namespace APIs for Redshift Serverless namespaces when opting in to this feature.

We are adding a new boolean parameter "ManageAdminPassword" to allow customers to opt in to this feature and another parameter "AdminPasswordSecretKmsKeyId" allows customers to specify the key ID of the KMS key in the customer account which will be used to encrypt the namespace's secret. These parameters can be used while setting CreateNamespaceRequest and UpdateNamespaceRequest. The response of these requests will return the "AdminPasswordSecretArn" when the namespace is opted in to this feature.